backend/requirements.txt
# API Framework
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
starlette==0.36.3
fastapi-cache2==0.2.1
orjson==3.9.15
asgi-lifespan==2.1.0

# Database
neo4j==5.18.0
redis==7.2.4
motor==3.3.2
asyncpg==0.29.0
alembic==1.13.1
sqlalchemy==2.0.27

# AI and Machine Learning
langchain==0.1.11
langchain-openai==0.0.8
langchain-anthropic==0.0.5
langchain-community==0.0.19
openai==1.12.0
anthropic==0.18.1
perplexity-python==0.1.1
numpy==1.26.4
scikit-learn==1.4.1.post1
sentence-transformers==2.5.1
torch==2.2.0
transformers==4.38.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
httpx==0.27.0
pytest-env==1.1.3
coverage==7.4.1
pytest-benchmark==4.0.0
locust==2.24.0
faker==22.6.0
hypothesis==6.98.9

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
cryptography==42.0.5
python-multipart==0.0.9
pyjwt==2.8.0
authlib==1.3.0
itsdangerous==2.1.2
oauthlib==3.2.2
python-keycloak==3.9.1

# Monitoring and Logging
prometheus-client==0.20.0
python-json-logger==2.0.7
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-instrumentation-neo4j==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-logging==0.44b0
elastic-apm==6.19.0
sentry-sdk==1.40.4
ddtrace==2.3.1

# Utilities
pandas==2.2.1
email-validator==2.1.1
aiohttp==3.9.3
python-dateutil==2.8.2
pytz==2024.1
ujson==5.9.0
pydantic-extra-types==2.5.0
rich==13.7.0
structlog==24.1.0
tenacity==8.2.3
cachetools==5.3.2
python-slugify==8.0.4
pyyaml==6.0.1

# Development Tools
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
autoflake==2.3.0
pylint==3.1.0
pre-commit==3.6.2
bandit==1.7.7
safety==2.3.5
ruff==0.2.2
pytype==2024.2.13
vulture==2.11

# Added for additional ML capabilities
transformers==4.38.1 